+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @tailor-platform/sdk
 
+## 0.15.1
+
+### Patch Changes
+
+- [#155](https://github.com/tailor-platform/sdk/pull/155) [`a65171a`](https://github.com/tailor-platform/sdk/commit/a65171af65109c08023a6b1a42683645c4e2675e) Thanks [@toiroakr](https://github.com/toiroakr)! - feat: add type constraints to workflow job body functions
+  - Input type: Must be JSON-compatible (no Date/toJSON objects). Interfaces are now supported.
+  - Output type: Allows Jsonifiable values (including Date with toJSON), undefined, and void
+  - Trigger return type: Returns `Jsonify<Output>` - Date becomes string after JSON serialization
+  - Added `JsonCompatible<T>` helper type to support TypeScript interfaces as input types
+  - TailorDB timestamp fields now return `Date` objects instead of ISO strings
+
+## 0.15.0
+
+### Minor Changes
+
+- [#145](https://github.com/tailor-platform/sdk/pull/145) [`8e4de4e`](https://github.com/tailor-platform/sdk/commit/8e4de4efa4e46117869ad32082cc5b54be2250f8) Thanks [@riku99](https://github.com/riku99)! - feat!: Change db/t enum arguments from variadic to array
+
 ## 0.14.3
 
 ### Patch Changes